L2-226: Refactor errors
