+++ src/types/ledger.ts
@@ -1,5 +1,28 @@
 import { Agent } from "@dfinity/agent";
 import { Principal } from "@dfinity/principal";
+import { ICP } from "../icp";
+import { BlockHeight } from "./common";
+
+export type TransferError =
+  | InvalidSender
+  | InsufficientFunds
+  | TxTooOld
+  | TxCreatedInFuture
+  | TxDuplicate
+  | BadFee;
+
+export class InvalidSender {}
+export class BadFee {}
+export class InsufficientFunds {
+  constructor(public readonly balance: ICP) {}
+}
+export class TxTooOld {
+  constructor(public readonly allowed_window_secs: number) {}
+}
+export class TxCreatedInFuture {}
+export class TxDuplicate {
+  constructor(public readonly duplicateOf: BlockHeight) {}
+}
 
 export type Fetcher = (
   agent: Agent,
